Generic DAO
The initial commit of the project introduced a persistence mechanism
that mapped fields of objects to database columns manually. This patch
changes that to store the objects using protocol buffers JSON
serialization iside JSONB fields in the database. That makes it possible
to use a single generic DAO for all the objects, and simplifies both the
database access code and the servers, as there is no need to manually
map between both.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>