Add database listener and notifier
This patch adds utilities to simplify the usage of PostgreSQL's
LISTEN/NOTIFIY mechanism. The `Listener` object takes care of keeping
the connection to the database open, even if the database is restarted.

Both the `Listener` and `Notifier` object use protocol buffers messages,
and they take care of encoding them so that they can be sent as payload
of the LISTEN/NOTIFY mechanism.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>