Add `dev listen` utility
This patch adds a new `dev listen` sub-command to the
`fulfillment-service` binary that listens for notifications published
using the PostgreSQL `NOTIFY` mechanism. It is intended to help
developers debug the messages produced by this mechanism.

To use it run the command providing the database URL and the channel
name, for example:

```shell
$ ./fulfillment-service dev listen \
--db-url postgresql://user:pass@localhost/db \
--channel events
```

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>