Add cluster reconciler
This adds a new reconciler that watches the _Cluster_ objects and
updates the Kubernetes _ClusterOrder_ accordinlgy. In particular it
copies the definitions of the node sets into the `nodeRequests` spec of
the _ClusterOrder_.

In order to do that change the patch also moves the cache of hub
connections to a separate component, so that it can be used by both the
cluster order and cluster reconcilers.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>