Always read token from file
Currently when the gRPC client is configured to read the token from a
file it will only read it when the service starts. That means that when
the token expires the server will start failing and will need to be
restarted. To avoid that this patch changes the gRPC client so that it
will always read the file.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>