Compare objects without removing metadata
When an object is updated we need to compare it to the current version
to avoid sending update notifications when there were no changes. This
comparison needs to ignore the creation and deletion timestamps. To do
so currently we remove the metadata from the objects before the
comparison, then we do the comparison, and finally we restore the
metadata. This will be difficult in the future when we introduce new
metadata fields that need to be compared. To make that easier this patch
changes the code so that instead of removing the metadata it explicitly
ignores the creation and deletion timestamps.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>