Set default node sets from template
This patch changes the clusters server so that it sets the value of the
`spec.node_sets` of clusters from the template. This used to be part of
the operator, because the node sets where not part of the cluster order,
but now that `ClusterOrder` and `Cluster` have been merged the
fulfillment service can do it.

The patch also introduces validations to ensure that users can't create
node sets that aren't defined in the template, or use host classes that
don't match the template.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>