Fix JSON encoding of Any
The protobuf `Any` type should be converted to the corresponding JSON
type, but our custom encoder doesn't do that, instead it converts its
fields. For example, an object created like this:

```go
object, err := anypb.New(wrapperspb.Bool(true))
```

Results in the following JSON:

```json
{
  "type_url": "type.googleapis.com/google.protobuf.BoolValue"
  "value": "CAE=",
}
```

But it should instead produce this:

```json
{
  "@type": "type.googleapis.com/google.protobuf.BoolValue"
  "value": true
}
```

This patch fixes that issue modifying our encoder so that it calls the
protobuf one for well known types like `Any` and `Timestamp`.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>