Add support for installing CRDs in the integration tests cluster
This patch adds a new `AddCrdFile` method to the builder of the kind
cluster that can be used to add CRD files that will be installed in the
cluster. This is used to install the `ClusterOrder` and `HostedCluster`
CRDs that will be used (in other patches) to write tests that verify
that `ClusterOrder` objects are created correctly.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>