Store notifications payload in a table
Currently the payload of notifications is sent using the PostgreSQL
`NOTIFY` mechanism, but that mechanism has a limit of 8000 bytes, which
is very easy to exceed. For example, it is exceeded for hub
notifications that contain large Kubeconfigs. To address that issue this
patch adds an unique identifier to each notification, and stores the
payload in a new `notifications` table. The payload of the `NOTIFY`
mechanism will now be only the identifer, and the listener will fetch
the payload from the table.

Related: https://github.com/innabox/fulfillment-service/issues/82
Related: https://www.postgresql.org/docs/current/sql-notify.html
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>