Move validation logic to private server and implement delegation
Currently validation is implemented only in public servers. But it is
desirable to have them also in the private servers. This patch does that
by moving the validations to the private servers, and changing the
public servers so that they delegate almost everything to the private
servers. A side effect of that is that the generic private server no
longer needs to map private objects to private objects, and the other
way around, so it no longer needs to have two generic parameters.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>