Use time-sortable UUIDs
This patch changes the service so that it uses `UUID7` to generate
unique identifiers. The advantage of this is that these identifiers are
time-sortable.

The implementation uses the same Google UUID library that we used in the
past, only that it uses the `uuid.NewV7` method.

All the places in the project that generate UUIDs have been changed to
use the new `uuid.New` function, so that if we decide to use some other
mechanism to generate identifiers we will only have to change that
function.

Related: https://github.com/innabox/issues/issues/223
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>