Add metadata server
This patch adds the implementation of the metadata server. It currently
only returns the list of trusted token issuers. That list is taken from
the value of the new `--grpc-authn-trusted-token-issuers` command line
parameter.

Note that this command line option, and the resulting metadata, is
purely informative. The way to define the token issuers that are
actually trusted is to change the Authorino configuration. Future
patches will change that configuration to a Keycloak instance deployed
using the Helm chart in the `charts/keycloak` directory.

Related: https://github.com/innabox/fulfillment-api/pull/48
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>