Fix tenancy check in events server
Currently the events server uses the context passed by the PostgreSQL
event listener to run the code that filters events. This context doesn't
contain the subject of the user that is watching the events. As a result
that code crashes when trying to get events. To avoid that this patch
changes the subscription object so that it contains the subject of the
original user, and so that the code that evaluates the tenancy logic
explicitly creates a new context containing it.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>