Update to common 0.0.13
This patch updtes the server to use version 0.0.13 of the common
library. The more relevant change in that version of the library is the
addition of a token source that reads tokens from a file.

In addition this path also fixes an issue with integration tests: they
where using the `...:latest` image of the service, but it has been
renamed to `...:main` some time ago. To avoid that category of issues
completely this patch changes the integration tests so that they build
and user their own image, using a timestamp as the tag.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>