Retry integration tests actions
The integration tests do things like patching an object and then
expecting that it will be inmediately available with the new state. But
this may not be true due to the use of caches in the API server and in
our controller. To avoid that this patch changes those parts of the
integration tests to retry.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>