Returns sets from attribution and tenancy logic
Currently the attributionn and tenancy logic return slices of strings
containing the identifiers of the tenants, and special meaning is given
to the empty slice: it means that tenancy is disabled. That is fragile
because it is very easy to confuse that with an empty set of tenants. To
avoid that this patch changes the tenancy logic to use a new `Set` type
that can explicitly represent the set of all tenants, a finite set of
tenants and the empty set of tenants.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>